api/requirements.txt
Flask
Flask-Cache
Flask-Migrate
Flask-Moment
Flask-OIDC
Flask-RESTplus
Flask-Script
Flask-SQLAlchemy
eventlet
flask-caching
flask-cors
flask_marshmallow
flask-socketio
gunicorn
kombu
marshmallow
marshmallow-sqlalchemy
mysqlclient
python-jose
